Restriucture webtool to work under custom command
