+++ surfiamviz/webutils/createtab.py
@@ -1,0 +1,8 @@
+"""Create."""
+
+import streamlit as st
+
+
+def create():
+    """Create own graphs tab."""
+    st.title("Create your own SRAM organisation graph.")
